DAO для работы с анализами

+++ src/main/java/com/dolzhenko_m_s/doctorhelpsystem/models/AnalysisResult.java
@@ -13,7 +13,7 @@ public class AnalysisResult {
     @Setter(AccessLevel.NONE)
     private long id;
 
-    private int patientId;
+    private long patientId;
 
     private @Nullable String analysisName;
